Fix typos
Currently, the Go and Node.js links are old URLs. For the Node.js link, this is okay because it seems like the repo was renamed and GitHub automatically redirects to the new path. However, for the Go link, the original repo was archived and a new one was created, so GitHub does not redirect to the new path.

In addition, the Go code sample was missing an ending parenthesis, using smart instead of dumb quotes, and was missing the package name. This makes it so users can't just copy paste the text.

This change updates the URLs and fixes the syntax errors.